# apps/exposed-api/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiResponse": {
				"properties": {
					"resCode": {
						"type": "number",
						"format": "double"
					},
					"resData": {
						"$ref": "#/components/schemas/ResponsePayload"
					}
				},
				"required": [
					"resCode",
					"resData"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"_36_Enums.Status": {
				"type": "string",
				"enum": [
					"Active",
					"Inactive",
					"Restricted"
				]
			},
			"Status": {
				"$ref": "#/components/schemas/_36_Enums.Status"
			},
			"UpsertLineups": {
				"properties": {
					"sportId": {
						"type": "number",
						"format": "double"
					},
					"lineupId": {
						"type": "number",
						"format": "double"
					},
					"challengeId": {
						"type": "number",
						"format": "double"
					},
					"challengeResultId": {
						"type": "number",
						"format": "double"
					},
					"pickTeamId": {
						"type": "string"
					},
					"spreadPoints": {
						"type": "number",
						"format": "double"
					},
					"gameId": {
						"type": "number",
						"format": "double"
					},
					"pickStatus": {
						"$ref": "#/components/schemas/Status"
					}
				},
				"required": [
					"sportId",
					"challengeId",
					"challengeResultId",
					"spreadPoints",
					"gameId",
					"pickStatus"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdateTiebreaker": {
				"properties": {
					"challengeResultId": {
						"type": "number",
						"format": "double"
					},
					"tiebreaker": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"challengeResultId",
					"tiebreaker"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "exposed-api",
		"version": "1.1.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/v1/sports/challenges": {
			"post": {
				"operationId": "CreateChallenge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Sports"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateChallenge"
							}
						}
					}
				}
			}
		},
		"/v1/sports/challenges/{challengeId}": {
			"post": {
				"operationId": "JoinChallenge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Sports"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "challengeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/JoinChallenge"
							}
						}
					}
				}
			}
		},
		"/v1/sports/lineups": {
			"patch": {
				"operationId": "UpsertLineups",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Sports"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpsertLineups"
							}
						}
					}
				}
			}
		},
...
		"/v1/sports/default-odds": {
			"get": {
				"operationId": "GetDefaultOdds",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Sports"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/sports/join-challenge-form": {
			"post": {
				"operationId": "JoinChallengeForm",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Sports"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/JoinChallengeForm"
							}
						}
					}
				}
			}
		},
		"/v1/sports/tiebreaker": {
			"patch": {
				"operationId": "UpdateTiebreaker",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Sports"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTiebreaker"
							}
						}
					}
				}
			}
		},
		"/v1/sports/challenges/type": {
			"patch": {
				"operationId": "UpdateChallengeType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Sports"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateChallengeType"
							}
						}
					}
				}
			}
		},
		"/v1/sports/challenges/smartcontract": {
			"patch": {
				"operationId": "SmartContractResponse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Sports"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SmartContractResponse"
							}
						}
					}
				}
			}
		},
		"/v1/sports/challenges/cancel/{challengeId}": {
			"patch": {
				"operationId": "CancelChallenge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Sports"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "challengeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/sports/challenges/favorites": {
			"get": {
				"operationId": "GetFavorites",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Sports"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateFavorites",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Sports"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpsertFavorites"
							}
						}
					}
				}
			}
		},
		"/v1/sports/potential-returns": {
			"get": {
				"operationId": "GetPotentialReturns",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Sports"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/sports/challenges/upstake": {
			"patch": {
				"operationId": "UpstakeTokenQty",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpstakeTokenQty"
							}
						}
					}
				}
			}
		},
		"/v1/sports/challenges/{challengeId}/redis": {
			"delete": {
				"operationId": "RemoveJoinRedisKey",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "challengeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/sports/leagues": {
			"get": {
				"operationId": "GetLeagues",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Sports"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/challenges/{userId}": {
			"get": {
				"operationId": "GetUserChallenges",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8000",
			"description": "Local development server"
		}
	]
}